requirements.txt
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.1
django-heroku==0.3.1
djangorestframework==3.8.2
gunicorn==19.9.0
heroku==0.1.4
idna==2.7
psycopg2==2.7.5
python-dateutil==1.5
pytz==2018.5
requests==2.19.1
urllib3==1.23
whitenoise==4.1
pipenv==2018.7.1
